Revert "inline format args to please our pedantic overlords"
This reverts commit 8ba7d1994672e3f18a0720569baf9c2997331a9a.

+++ src/v4l2/api.rs
@@ -47,7 +47,7 @@ mod detail {
             prot,
             flags,
             fd,
-            offset,
+            offset as i64,
         )
     }
     pub unsafe fn munmap(start: *mut std::os::raw::c_void, length: usize) -> std::os::raw::c_int {
